requirements.txt
# TenxsomAI Production System Requirements

# Core Python packages
python-dotenv>=1.0.0
psutil>=5.9.0

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Google Cloud Tasks (Primary Queue System)
google-cloud-tasks>=2.19.0
cloud-sql-python-connector>=1.4.0
google-cloud-storage>=2.10.0
google-cloud-logging>=3.8.0
google-cloud-monitoring>=2.16.0

# Google Cloud and AI
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# YouTube API
google-api-python-client>=2.100.0
googleapis-common-protos>=1.60.0

# HTTP and API clients
requests>=2.31.0
aiohttp>=3.8.5
httpx>=0.24.1

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Telegram Bot Integration
python-telegram-bot>=20.5

# Data Processing
pandas>=2.0.3
numpy>=1.24.3

# Monitoring and Logging
structlog>=23.1.0
colorama>=0.4.6

# JSON and Configuration
pydantic>=2.3.0
jsonschema>=4.19.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.1